add MirRef_PeelIndex intrinsic
This operation removes an outermost Index_RefPath, returning both a
reference to the enclosing MirVector and the index of the input within
that vector.  This operation has no analogue in ordinary Rust, but is
useful in implementing certain overrides (particularly `reallocate`,
which takes a pointer to the first element of a vector but needs to
manipulate the vector as a whole).