Generalize asMemAllocationArrayStore to symbolic size allocations. (#507)
This change enables `asMemAllocationArrayStore` to work on symbolic size allocations and writes in the simple case when the allocation size and the write size are determined equal by `bvEq`. This is sufficient to handle the combination of `crucible_symbolic_alloc` and `crucible_array_prefix_points_to` when the size of the allocation and the size of the points-to are the same SAWCore term. Deciding at What4 level that the array write covers the entire allocation prevents a possible ite explosion when reading. Supporting array writes that partially cover the allocation is more complicated, and may be addressed in the future as needed.