Remove the `ReadMem` monad accounting that tried to keep track of which symbolic writes might have satisfied a load request.
I don't think it worked correctly with the caching mechanism,
and I think that memory concretization yields better results,
in terms of error comprehensibility.

In addition, play a trick to convince What4 to retain
the `NoSatisfyingWrite` annotation.  Previously, the assertion
would be quickly discarded if combined with, e.g., path conditions.
By making the final fallthrough case a fresh boolean variable (instead
of a false constant) it will not be simplified away, but will still
cause a verification failure in the correct conditions.