Initial version of a `crucible-wasm` package.
This package currently is capable of loading WebAssembly modules
and simulating a subset of concrte programs that use integer values,
control flow and memory. There is a very basic `crux` driver layer
that loads scripts from the `haskell-wasm` test suite and executes
them (only a subset of the assertions in those scripts are supported,
however).

There is plenty of work still to be done, but the existing subset
will allow some nontrivial programs to be simulated, and a bit more
crux infrastructure support would allow serious symbolic simulation
testing to begin.