[crux-llvm] Add -Wcompat and remove profiling flags from cabal config.
The profiling operations were removed because if crux-llvm is used as
a library elsewhere this would cause potentially unwanted cluttering
of the profile information for the downstream package.

Per the Cabal
documentation (https://cabal.readthedocs.io/en/3.4/cabal-package.html#pkg-field-ghc-prof-options),
the preferred way to handle this is to manually specify the
`--profiling-detail` or `--library-profiling-detail` command-line
options if a build with profiling detail from this package is desired.