[crucible-llvm] use local-relative path from .bc file, not absolute
Using absolute paths obtained from the compiled .bc file will both (a)
leak information about where the working tree was located and (b) be
irrelevant changes in golden-testing output that will cause test
failures.

The main advantage of the absolute path was that tools, editors, and
IDE's (e.g. emacs) can automatically load and move to the error point
in the associated file when an unsat counterexample is obtained.  An
alternative approach for this is to emit a line like:

  Entering directory `path/to/stuff'

at the beginning of the compilation/build process for emacs (or
whatever else is appropriate to the tool at hand).  If this change is
highly problematic, alternatives can be explored.