[crux-llvm] Add CI_TEST_LEVEL to test and github action CI config.
A CI_TEST_LEVEL of 0 (the default if unspecified) will run a set of
tests expected to complete in under a minute, suitable for a normal
development environment.

A CI_TEST_LEVEL of 1 (specified for github actions) will run the full
set of tests, even those that can take an extended period of time.