uc-crux-llvm: Improve type translation
Previously, the declaration for each function was separately translated into
`FullType`s, now all declarations are done at once. This has a few benefits:

- Efficiency: there's no duplicated work when exploring multiple functions
- Parsimony: The ModuleTypes abstraction applies best at the module level
- Functionality: The big feature introduced here is `findFun`/`CFGWithTypes`:
  the `FullType`s for a declaration can be looked up in the `ModuleContext`.
  One spot this will be useful is in generating return values from skipped
  functions.

Additionally, this patch introduces support for variable-arity function types.
This was necessary because the default is now to translate *all* function types
in a module, which would fail on any module with a varargs function if these
weren't supported.