crucible-jvm: Add permission bit for writability to each static field.
The permission bit is asserted to be true on every putstatic instruction.

This feature will allow SAW to use method specs that do not explicitly
specify a final value for each static field; by making those fields
read-only, we can be sure that the values did not change, without having
to explicitly check that the final value is equal to the initial value.