crucible-llvm: Only consult default address space for pointer size
LLVM 10 and later reserve multiple address spaces (with different sizes!) in
the data layout. Since `crucible-llvm` is only equipped to deal with a single
pointer size throughout a program at the moment, we now deal with this by only
using the default address space in the data layout to determine what the
pointer size will be. This means that any programs that actually _use_
multiple address spaces of different pointer sizes are unlikely to work
in `crucible-llvm`, but we do not aim to support such programs yet.

While I was in town, I took the liberty of documenting the assumptions that
`crucible-llvm` makes with respect to address spaces in
`crucible-llvm/doc/limitations.md`.

Fixes #737.