crux-llvm: Search in more places for the default lib-dir (#746)
Previously, `crux-llvm` was relying entirely on `getDataDir` to find the
location of the default `lib-dir`. However, this proves fragile enough that
we should consult other locations besides it. (See the discussion in #731.)
The most obvious location to search is in a directory relative to the
`crux-llvm` binary itself, which serves the common use case of invoking a
`crux-llvm` bindist.

Fixes #731.