use data types for log messages across all projects
This commit builds on top of the recent Lumberjack integration.  Instead of
having formatting and tagging done at call sites to the logger, this commit
proposes an alternative design where all log messages are part of a data type,
so that the handlers get to decide which messages they want to print, as well
as what the best formatting for their output is.

This allows to easily change the handling of all messages so as to accomodate
non-standard outputs, such as machine-readable serialization of outputs for IDE
backends.

Since a lot of projects rely on Crux but also add their own output, the
'CruxLogMessage' data type allows for an extension point where projects may add
their own data type extension.  Handlers for such extension log messages are
passed via implicit parameters, a feature already used pervasively for logging
decisions in the code base.