Add new simulator states for assertions and assumptions.
These states make more explicit the steps of program execution
that add assumptions or assertions to the logical state of the program
so that they can be intercepted by execution features.

Currently, only explicit `Assert` and `Assume` statements are using
these new states.  Some rather extensive refactoring will be required
to migrate to this new system.