Refactor "mux" functions to handle mux failures differently.
Instead of immediately making an assertion via the `sym` when values
cannot be muxed, we modify the monads in which muxing is happening
to carry along that failure case.  This pushes the assertions
closer to the simulation loop, where we will be able to use the
new assertion states instead.