Add "event" tracking to the assumptions data structures.
This allows us to track program events alongside the existing
assumption tracking. Using this, we can reconstruct a linear sequences
of events that occured along the program run when we find
counterexamples to proof goals.

Initially, the only event we track is the creation of user-requested
symbolic variables, but others are planned.  By tracking symbolic
variables this way, we can accurately reconstruct a linear sequence
of values to replay, which allows us to fix #587.