Rearrange some code relating to crux model views.
In crux-llvm, we need to make sure that declarations are emitted for each of
the types so the counterexample code will compile.  Previously,
this was handed by prepopulating the Model map, but this is fragile
and mixes crux-llvm concerns into crux proper.  Here, we split this
concern out, and push the handling of this issue into crux-llvm,
where it belongs.