crux: log prover milestones
For vscode-crux-llvm, we would like to have a notification when Crux knows what
goals it is about to start solving, and whenever each goal is started and
ended.

The goals are too verbose for displaying to the user, so they are not.  We
already have a progress bar system for the goals starting/ending, so these
messages are also not displayed to the user.

Because I did not want to build on the assumption that a goal ending was
necessarily the last started goal, I also made changes so that the ending of a
goal repeats the goal number.  This should be more future-proof in case we ever
process goals in parallel.  It can also help IDEs be a little more stateless.

Also, my apologies for inadvertently introducing 'StartedGoal' in a previous
commit where it was unused.