symio: Improve the filesystem initialization API
Make the initial file contents a separate type and introduce a(n optional)
helper function to create initial filesystem contents based on an on-disk format
consisting of concrete files and a symbolic manifest.

More documentation for that format is coming shortly.  It is also rudimentary
right now; it will be extended as-needed.