Thread the symbolic IO setup code through to crux-llvm
This adds an option to crux-llvm to specify an initial filesystem.  It calls the
function provided by SymIO.Loader to load content from the filesystem.  It
provides sufficient infrastructure for the tests to pass if a symbolic
filesystem is not provided.  It also provides the mechanisms to register all of
the initial file descriptors for standard IO.

Note that the standard IO setup likely does not work because there are no files
inserted for standard output/standard error when they are requested (and thus
the call to `openFile` will probably generate an assertion failure due to the
files being missing).