Revert "Add a case to `unpackMemValue`"
This reverts commit fc3dde845e3569195c27770a2a085b7930e1852f.

We can fix this in a different way that is more consistent with the rest of the
design of crucible-llvm. The call site that required this needs
bitvectors (because it is creating an SMT array), but we can do the necessary
load at LLVMPointer type and then project out bitvectors at the call site.