Modify how we load chunks from memory in the LLVM frontend
We were previously loading bytes as bitvectors. This triggered an error in the
LLVM memory model due to that exact case being unsupported. The choice was
apparently intentional to catch errors in the LLVM translation. We can work
around it here by loading at the supported LLVMPointer type and then projecting
locally to a bitvector.  This will always be safe because the width of the
operation is not pointer sized.