crux-llvm: add TypeScript instances and executable to generate definitions
In order to ease the synchronization of the VSCode plug-in with changes made to
the logging infrastructures in crux and crux-llvm, it helps to have the Haskell
code generate TypeScript type definitions of the logging messages tags and the
type of their payloads.

While this can be done somewhat manually, the aeson-typescript automates the
tedium of it.  This commits uses it to derive 'TypeScript' instances for the
relevant log messages, and creates an executable 'typescript-of-crux-llvm' that
outputs the type definitions needed to interface against 'crux-llvm'.