Symbolic I/O (#788)
Add support for symbolic I/O in Crucible.

This support has three main parts:
1. The core implementation of the data structure for representing symbolic files (crucible-symio)
2. A wrapper in crucible-llvm adapting the low-level primitives to the POSIX primitives as they appear in C
3. Integration in crux-llvm (including an interface for specifying initial symbolic filesystems)

This feature is designed to enable users of crucible to verify properties of programs that interact with the system (via file I/O), with the specific goal of supporting programs with configuration files. The standard I/O streams have been hooked up to their POSIX file descriptors in crux-llvm; see the README for the current limitations.

The key observation behind this functionality is that files are essentially memory regions with a file pointer that indicates where the next read/write will occur. The data structure backing crucible-symio is a byte-oriented memory model with a write log, much like the LLVM memory model (without extra layers for representing LLVM types).

Note that symbolic I/O is still experimental and expected to change over time.

Co-authored-by: Daniel Matichuk <dmatichuk@galois.com>