Broaden scope of read-unwritten-memory to lax-loads-and-stores
This renamed the `read-unwritten-memory` Crux option to `lax-loads-and-stores`,
reflecting its new purpose of being a catch-all option for relaxing various
validity checks related to loads and stores. This is meant for SV-COMP mode,
which has a different view of fatal errors than Crucible does.

This addresses some of the goals of #783. It does not address the goal of
designing a generic taxonomy of errors.