Overhaul crux-llvm-svcomp to fit in with SV-COMP's benchmark harness
This rewrites the code in `crux-llvm-svcomp`. Rather than trying to emulate
the benchmarking code that SV-COMP uses to grade a competition
participant and run many input files at a time, `crux-llvm-svcomp` now takes
a single input program and judges whether it meets a particular specification.
This fits in better with the way that SV-COMP's own benchmark harness works.
(I will add notes on how to set up that benchmark harness in a subsequent
commit.) Many of the options in `Crux.SVCOMP` and the logging settings in
`SVComp.Log` are therefore no longer relevant and have been removed.

In addition, `crux-llvm-svcomp` now produces witness automatons, which is
an output requirement started in recent SV-COMPs. The details of the automaton
file format can be found in `Crux.SVCOMP.Witness`.