uc-crux-llvm: Improve interface to LLVM modules
The primary purpose of this change is to move generic types relating to
LLVM modules out of UCCrux.LLVM.FullType.Translation, and into a new
module UCCrux.LLVM.Module. This makes it easier to maintain the
invariants associated with these types, because less code has access to
their privileged constructors.

Additionally, this change makes it more obvious that the types and
functions involved aren't really specific to their application inside
UC-Crux-LLVM, and can be reused in other contexts.