Add overrides for `hton{s,l}` and `ntoh{s,l}`
These are identical to the overrides for `llvm.bswap.*`, except that they
only swap bytes if being run on a little-endian machine. I have added a test
case which ensures that when these functions are run on a little-endian
architecture, their behavior corresponds to that of `bswap_16` and `bswap_32`
from `byteswap.h`.

Fixes #840.