Add override for llvm.ubsantrap()
This is a new override introduced in LLVM 12 that acts like `llvm.trap()`, but
with an argument indicating which form of undefined behavior was trapped. As a
result, `crucible-llvm`'s override can reuse the existing logic for the
`llvm.trap()` override. Ideally, we would do something more intelligent with
the argument value, but doing so would likely require fixing #368 first. Until
then, this simpler approach is good enough.

Fixes #845.