uc-crux-llvm: Use explicit return statement in linked_list_sum.c
This is needed when running the `uc-crux-llvm` test suite with LLVM 12, as
Clang 12 more aggressively optimizes away the body of `linked_list_sum` due to
it not ending with a `return` statement, despite its type claiming to return an
`int`, resulting in a simple function that returns `undef` in LLVM bitcode. I
believe the lack of a `return` statement is due to an oversight, since
`mututally_recursive_linked_list_sum.c` does have a `return` statement, so the
simplest fix appears to be to add one to `linked_list_sum.c` as well.