Crudely translate `freeze` instructions
We handle `freeze` instructions in `crucible-llvm` by simply passing the
argument value unchanged. This isn't the intended semantics of `freeze`
(see https://releases.llvm.org/12.0.0/docs/LangRef.html#id323), where `undef`
or `poison` arguments should result in `freeze` returning an arbitrary value.
Unfortunately, Crucible isn't quite equipped to accurately determine if an
given value is `undef` or `poison` in all situations (see #366), so this is the
best that we can do for now.