uc-crux-llvm: Overrides that check inferred contracts
'Check' overrides take the result of contract inference on a function f
(namely, 'Constraints') and create an override that has the same
signature as f, and when called will

- Create a bunch of predicates that represent whether or not the
  constraints were violated at this call
- Stash them away in an IORef
- Call the original implementation of f

The hope is that this will be useful for refining overly pessimistic
(sufficient but unnecessary) preconditions in inferred contracts, or for
indicating the presence of bugs (when the contract really *should*
hold).

This is not currently integrated with other functionality (i.e., it's
not reachable from the CLI), but is instead a minimal prototype.