Add overrides for llvm.copysign.{f32,f64}
LLVM 12 is much more eager to simplify certain expressions to calls
to `llvm.copysign.f*`, as illustrated in
https://github.com/GaloisInc/crucible/issues/187#issuecomment-917047901.
It proves relatively straightforward to just add overrides for the
`llvm.copysign.f{32,64}` instrinsics, so this patch accomplishes just that.
The one caveat is that we cannot accurately represent the semantics of
`llvm.copysign.f*` with respect to NaN arguments, since What4 does not
possess a way to distinguish the sign bits of NaN values. I decided not to
let perfect be the enemy of good, however, so I simply made a note of this
limitation in the `crucible-llvm/doc/limitations.md` file.

Addresses one bullet point in #187.