Add overrides for llvm.abs.* and friends
`llvm.abs.*` is an intrinsic that was added in LLVM 12 that behaves quite
similarly to C's `abs` family of functions. They are similar enough that I
opted to combine the logic for all of these overrides into a single `callAbs`
function. I use a `CheckAbsIntMin` data type to smooth over the minor API
differences, such as what to do when the argument is `INT_MIN` (in `abs`,
it is UB, while in `llvm.abs.*`, it is poison).

Fixes #847.