crux-llvm: Support .at-least-clang<VER>.good files in test suite
We have the ability have test cases that should be run on specific
versions of Clang (e.g., `test.clang12.good`), but we also ought to have a way
to run tests on a _minimum_ version of Clang. This patch adds an
`at-least-clang<VER>` modifier which does just that. This will be put to use
in a subsequent commit.