uc-crux-llvm: Refactor handling of StorageTypes
'toStorableType' only has one failure mode that isn't relevant to
uc-crux-llvm (which doesn't support/use metadata types), so wrap it in a
non-partial function (that panics) in its own module to promote reuse. The
places that handled the error case already paniced, so the error handling hasn't
fundamentally changed, it's just more clear what's at fault if a panic arises.

Also, the function that turns a Constraint into a Pred is similarly moved into
its own module to promote reuse (and there's a planned change that will use
this).