Add overrides for some special floating-point functions
This patch:

1. Bumps the `what4` submodule to bring in the changes from
   GaloisInc/what4#160, which adds support for special floating-point
   operations (called `SpecialFunction`s).
2. Adds overrides for the `libc` functions and LLVM intrinsics corresponding
   to each `SpecialFunction`. While I was town, I also added overrides for
   `sqrt`, which does have some support in SMT solvers (and hence its own
   method in `IsInterpretedFloatExprBuilder`).

I have added two basic test cases to test that each override has code coverage.
The `special-functions` test case checks the `libc` overrides, and the
`special-functions-fast-math` test case checks the LLVM intrinsics, which seem
to be used when things like `-ffast-math` are enabled.

Because this implements support for `llvm.sin.*`, `llvm.cos.*`, and
`llvm.sqrt.*`, this ticks off some boxes in #187.