fix no-colors for help/version (#878)
This required a restructuring of Crux options where the options relating to output are a separate record, of which just the color options are yet another sub-record.

Generic lenses can be used for nested record updates, while the accessors have remained unchanged. Client code is encouraged (but not required) to use the lenses for brevity in updating the sub-records.