uc-crux-llvm: Begin using the new Crux 'resultHook'
The goal is to expose a lot of the same flexibility that exists in
Crux's 'SimulatorCallbacks' through the UC-Crux API, while maintaining
the current behavior along the way.

This commit begins the process by allowing a configurable result hook in
'simulateLLVM', this will eventually be exposed through 'runSimulator'
and its callers just like what was done for registering overrides
('CreateOverrideFn').