crucible-llvm: Simplify 'MemoryErrorReason'
There was only one constructor of this type that depended on type
parameters 'sym' and 'w', and that was a pointer that is duplicated in
the 'MemoryOp' that accompanies each 'MemoryErrorReason'.

Also, create a new error datatype instead of using a pretty-printed doc.

In addition to making 'MemoryError' more normalized, 'MemoryErrorReason'
is now friendlier to programmatic use. In particular, it's orderable.