L.C.LLVM.SymIO: export more override functionality
One can use the already exported `LLVMOverride`s with `llvmOverride_def`, but
this has the drawback of forcing the `ext` to be `LLVM`. However, there is
nothing about the underlying `OverrideSim`s for these overrides that inherently
requires the `ext` to be `LLVM`. This patch exports the underlying
`OverrideSim`s, generalizing their `ext`s in the process for the benefit of
consumers who need to use them in extensions besides `LLVM`.