`crucible-llvm`: Control granularity of reading uninitialized memory (#906)
* Revert "Use readUnwrittenMemory in readMem'"

This reverts commit 819adc8b14457b8907082f139749cf7791bcdfa2.

* crucible-llvm: Control granularity of reading uninitialized memory

Previously, `laxLoadsAndStores` would make it so that all reads from
uninitialized memory would return a new, fresh value. This isn't always what
you want, however, as this implies that consecutive reads from the same,
uninititalized variable would return different results. This patch adds another
knob in the form of the `indeterminateLoadBehavior` field of `MemOptions`,
which has two settings:

* `unstable-symbolic`: The previous behavior, in which each read from
  uninitialized memory will return a fresh symbolic value each time.
* `stable-symbolic` (now the default): After memory is allocated (be it
  through `alloca`, `malloc`, `calloc`, etc.), it is written with a fresh
  symbolic value. This makes it so that reading from the same section of
  uninitialized memory multiple times will always return the same symbolic
  value.

Note that `indeterminateLoadBehavior` only takes effect when `laxLoadsAndStores`
is enabled.

Fixes #844.

* Only apply the `stable-symbolic` semantics for heap allocations in
`doMalloc`.

This avoids a bug that was causing global variables to be incorrectly
initialized. The ultimate source of that bug still needs to be fixed,
but this avoids the bug for now, and is also the right thing to do.

* Remove a fall through case in `writeMemWithAllocationCheck`
that was sometimes causing memory writes to be silently
discarded.

In particular, writing an `LLVMZero` value into an SMT array
backed location would have no effect.  Unhandled cases in this
code will now panic instead of silently giving incorrect answers.

* Fix a bug in global initialization that could incorrectly "undo" some
global variable setup if there is a chain of global variables that
refer to others via constant expressions.

* Update crucible-llvm test suite

* Include globals in T844 test cases

* Handle "extern" global variables in the lax-loads/stable-symbolic case.

This involves making a special case so that external globals recive
an initializing write with fresh sybmolic values at global population
time.

* Add some notes about deferencing pointers with lax-loads-and-stores

Co-authored-by: Rob Dockins <rdockins@galois.com>