crucible-llvm: Add overrides for the isnan family
This adds overrides for the `isnan`, `__isnan`, and `__isnanf` functions.
The first of these functions can appear as a function call in LLVM bitcode
when it is invoked as `(isnan)(x)` (as opposed to `isnan(x)`, which is treated
as a macro application), and the latter two functions appear in a handful of
SV-COMP benchmark programs.

Addresses two bullet points of #187.