Detangle the Crucible "backend" notion from the What4 expression builder.  There is now a new concept of the "simulator backend" which is separate from the older, and still existing concept of the "symbolic backend".
The What4 expression builder still plays the role of the "symbolic
backend" and continues to occupy the ubiquitous `sym` type
variable. It is required for tasks that involve constructing new
expressions and communicating with solvers. However, the new
"simulator backend" concept (occupying the `bak` type variable) is
responsible for maintining the assumptions/assertions corresponding to
control flow paths and simulation side conditions.

A `bak` value contains a `sym`, which may be used by multiple
different `bak` values, if required; previously the two were bound
together.

In practice, parts of the code that deal with control flow or want to
make assertions will need access to the `bak`, not just the `sym`.
This includes most implementations of simulator overrides, so this
change will necessitate quite a few downstream changes; however, the
majority of these should be quite mechanical.