Drop support for GHC 8.6 (#950)
Now that `crucible` supports GHC 9.0, it's about time to drop support for 8.6,
as we generally only support the previous three stable releases of GHC (in this
case, that would be 8.8, 8.10, and 9.0). As a result, lots of 8.6–specific
CPP can be removed. I have updated the version bounds of affected packages
accordingly.