Fix -Wnoncanonical-{monad,monoid}-instances warnings
GHC 9.2 adds `-Wnoncanonical-monad-instances` and
`-Wnoncanonical-monoid-instances` to `-Wall`, which warn whenever one has
explicit implementations of `return` or `mappend` that aren't simply
`return = pure` or `mappend = (<>)`. Since these are the default
implementations of `return` and `mappend` since `base-4.11`, the simplest
way to fix the warnings is to simply remove all explicit definitions of
`return` and `mappend` and rely on the defaults, which this patch accomplishes.