Fix -Wincomplete-uni-patterns warnings
GHC 9.2 now includes `-Wincomplete-uni-patterns` as a part of `-Wall`. As a
result, building `crucible-concurrency` and `crux-mir` with 9.2 uncovers some
previously undetected warnings. This patch fixes them by converting each
incomplete uni-pattern to a `case` that explicitly calls a bespoke `error`
function in the event of a match failure. This is no less partial than before,
but it avoids warnings and should provide a more specific error in the event
that these incomplete uni-patterns are triggered.