Pare down the number of CI configurations (#962)
* Pare down the number of CI configurations

Currently, we have a whopping 36 (!) different number of CI configurations.
This is mainly because we test every supported version of GHC across both
Linux and macOS, and in the case of `crux-llvm-build`, we even test multiple
versions of Linux. I believe this is a bit overkill, so this patch attempts to
trim down the number of configurations to something more manageable:

* For each project, we test the latest LTS release of Ubuntu on all supported
  versions of GHC.
* In addition, we also test the previous LTS release of Ubuntu on the
  `crux-llvm-build.yml` workflow, but only using the latest supported version
  of GHC.
* For each project, we test macOS only using the latest
  supported version of GHC. This is because the turnaround time for macOS
  builders on GitHub Actions is longer, so we try to avoid
  bottlenecking CI workflows on macOS builds.

  In the future, we wish to test Windows using the latest supported version
  of GHC as well.

I've written this up in `.github/workflows/README.md`.

* Only invoke Nix on Linux

It's only really needed on Linux anyway. In a future where we wish to test
Windows on CI, this will be required since Nix does not work on
Windows.