Remove unused submodules and optional-dependencies (#965)
Although the `cabal.project` file declares dependencies on `abcBridge`, `blt`,
`what4-abc`, and `what4-blt`, none of the projects in `crucible` actually
depend on them in practice. This patch removes those dependencies and the
associated submodules. This means fewer packages to build with running
`cabal build all`, which is all around a good thing.