uc-crux-llvm: New type for function postconditions
The Postcondition module declares a type that approximates a
specification of a function's postconditions. This is used in Classify
to infer such postconditions on return values, in the future it may be
used to infer information about clobbered memory.

This module supersedes the "Clobber" specs that were part of the Skip
override machinery. These weren't amenable to use-cases such as
inference.

Also implement some additional pretty-printing machinery.