crucible-syntax: Add forward declarations
This adds a `(declare ...)` form to the `crucible-syntax` language for
forward declarations. A forward declaration is like a function, but lacking a
body, and is useful for situations where one does not know what the
implementation of a function will be until after the `.cbl` file is parsed.

Implementation-wise, this is accomplished by adding a new
`parsedProgForwardDecs :: Map FunctionName SomeHandle` field to `ParsedProgram`
that maps the name of each forward declaration to its handle. It is the
responsibility of the caller to register each handle with a suitable `FnState`,
or else invoking the forward declaration will result in a simulation error.

I've added a basic test case to ensure that this plan works as intended. This
involved adding another field to `SimulateProgramHooks` (named
`resolveForwardDeclarationsHook`) to allow resolving forward declarations
before simulation begins.

Note that neither `crucibler` nor `cruces`, two of the current consumers of
`crucible-syntax` in the `crucible` repo, support forward declarations at the
moment, and trying to use forward declarations in the context of `crucibler` or
`cruces` will result in an error. One could imagine adding support, although
there would be some design questions to work out first.

Fixes #996.