Thread ?memOpts through to writeMemWithAllocationCheck
This will be required in a subsequent commit that changes the behavior of
`writeMemWithAllocationCheck` to behave differently depending on whether
`laxLoadsAndStores` is enabled or not.

This requires adding implicit constraints in several functions that did not
previously require them, so I have added mention of this in the affected
libraries' changelogs.